[tests] more logging around MicrosoftNetSdkDirectory (#2725)
Downstream in monodroid, we have a test failure:

	1) Failed : Xamarin.Android.Build.Tests.PackagingTest.NetStandardReferenceTest
	Microsoft.NET.Sdk not found.
	at Xamarin.Android.Build.Tests.PackagingTest.NetStandardReferenceTest ()

We likely introduced this in ef7a38d, but there isn't enough logging
to see what went wrong.  When I connected to the build machine, the
directory we are looking for looked correct.

I changed `PackagingTest.NetStandardReferenceTest()` to log the
directory it used, and added a `Console.WriteLine()` call to the
`Builder` constructor.