[Xamarin.ProjectTools] check for /msbuild/ in $(FrameworkSdkRoot)
Downstream in monodroid, we are still seeing a failure in
`NetStandardReferenceTest`:

    Using $(FrameworkSDKRoot): /Library/Frameworks/Mono.framework/Versions/5.18.0/lib/mono/
    [TESTLOG] Test NetStandardReferenceTest Complete
    ...
    Microsoft.NET.Sdk not found:

But if I test on my machine `$(FrameworkSDKRoot)` is:

    /Library/Frameworks/Mono.framework/Versions/5.18.0/lib/mono/msbuild/

Both the build machine and mine have the same version of Mono.

I think we should just make sure the path we are using includes
`/msbuild/` and append the directory if needed.

Other change:

* On Windows, it was always logging `Using $(FrameworkSDKRoot): `. So
  let's not log that if it is blank.