[installers] Share installer payload generation logic
Introduces a new `create-pkg` project next to `create-vsix` to be used
in a similar way to create a macOS installer for xamarin-android. The
`build-tools/installers/create-installers.targets` file has also been
added to keep track of all of the content we require in both our macOS
and Windows installers. These shared items, properties, and targets are
used by both installer projects to ensure uniformity between the two
installers we produce.

The `ConstructInstallerItems` target determines the majority of the files
which should end up in the installer. Behavior will vary depending on the
existance of certain files or the values of certain msbuild properties.
This was done intentionally to be able to support creating installers for
PR builds, open source packages, and commercial packages.

The `AndroidFirstFrameworkVersion`, `AndroidLatestStableFrameworkVersion`,
and `AndroidLatestFrameworkVersion` MSBuild properties (along with some
hardcoded values) are used to determine which API levels the installer
will support.

The `AndroidSupportedTargetJitAbi` MSBuild property is used to determine
which Android ABIs the installer will support.

The `IncludeMonoBundleComponents` MSBuild property is used to determine
whether or not the components which come from a full mono integration
build should be included in the installers.

The new `create-installers` make target will attempt to package all of
the relevant files for every ABI specified in the `ALL_JIT_ABIS` variable
in `build-tools/scripts/BuildEverything.mk`. This behavior can be
overridden by specifying an empty value for `_MSBUILD_ARGS` when calling
the `create-installers` make rule (as build.groovy does), or by providing
an alternate value for `/p:AndroidSupportedTargetJitAbis` when building
either installer directly from MSBuild.

The build.groovy script has been updated to handle creating different
installer types for PR builds and full builds. Our macOS PR builds will
now build a mxe-win flavored version of `libmono-android`, so that it can
be included in a vsix installer and used to extend test coverage for our
PR builds in the future. PR builds will also package a limited set of
Android ABI support as briefly mentioned above.

A new pkg preinstall script has also been added to fix an issue that
would result in old installations being left on disk. Now, the macOS
installer will attempt to remove the version currently symlinked to
`/Library/Frameworks/Xamarin.Android.framework/Versions/Current`
before running.

Fixes https://github.com/xamarin/xamarin-android/issues/2686.