[Xamarin.Android.Build.Tasks] Move the Designer Targets to Xamarin.Android
The Android Designer has a set of custom targets which it
uses to help prepare a project. This commit moves
them into the open source repository. The idea being that
we should be able to test the designer integration against
the open source repository rather than waiting for it to
be integrated into the consumer packages.