[monodroid] Hookup the jit_failed event as well
Refactor the events code a bit to avoid duplication

Note: I didn't see any failed events in the log. I hoped it might shed
light on parts like this one:

    JIT method  begin: (wrapper managed-to-native) Android.Runtime.Logger:__android_log_print (Android.Runtime.LogLevel,string,string,string,intptr) elapsed: 0s:139::103087
    JIT method   done: Android.Runtime.Logger:__android_log_print (Android.Runtime.LogLevel,string,string,string,intptr) elapsed: 0s:139::226837
    JIT method   done: (wrapper managed-to-native) Android.Runtime.Logger:__android_log_print (Android.Runtime.LogLevel,string,string,string,intptr) elapsed: 0s:139::232723
    JIT method   done: Android.Runtime.Logger:__android_log_print (Android.Runtime.LogLevel,string,string,string,intptr) elapsed: 0s:139::238712

I think it is still useful to have that event hooked up, so that we
see these in our `methods.txt` log file, when it happens. Looking at
the mono code, this event is raised in case of mono error and when
generic sharable method is not created properly.