[Xamarin.Android.Build.Tests] Dont auto clean up tests.
We have a problem in our tests when the build passes,
but the unit test fails. In this case the `ProjectBuilder`
is auto cleaning up the failed test build directory along
with ALL the logs.

So Lets NOT clean up by default. This does not mean that
we will leave files on disk for passed tests since the
`BaseTest.CleanupTest` [1] method is an nunit `TearDown` method.
So it runs at the end of the test and will clean up the
directory if the test passed.

[1] https://github.com/xamarin/xamarin-android/blob/master/src/Xamarin.Android.Build.Tasks/Tests/Xamarin.Android.Build.Tests/Utilities/BaseTest.cs#L376