[build] Drop requirement for MXE
MXE is a source-only distribution to build the MinGW Windows cross-compiler
on Unix systems, along with a number of libraries and tools. We use it to
build both Mono and Xamarin.Android for Windows.

The problem with MXE is that its build process is very fragile and also takes
a lot of time to complete. Additional issue is that the version of GCC it
builds is quite ancient (5.4 as opposed to the current 8.2) and does not support
the latest C and C++ standards which prevents us from using C++17 features in XA.

However, [Homebrew][0] now has a formula to install the latest MinGW toolchain with
GCC 8.2 which allows us to drop MXE. The downside of using this pre-packaged MinGW
(a very insignificant one) is that it does not come with any extra libraries or
software prebuilt for Windows and our build requires `zlib` which was previously built
as part of MXE. That downside is easily worked around, however - we simply use our
own Homebrew tap which has a formula that installs the Win32/Win64 zlib binary.

This commit drops MXE dependency in Xamarin.Android *only* - it does *not* remove this
dependency from Mono. Mono will now provision its own copy of MXE and use it to build.
This will change at some point when [this PR][1] makes its way into Mono and, subsequently,
to Xamarin.Android.

All the libraries we build for Windows that require zlib will link a *static* version
of the library instead of using the dynamic one. This is safer than using the DLL because
we control what version of zlib is used and we leave no potential for breakage should a
3rd party replace/remove/etc zlib1.dll from the target system.

[0]: https://brew.sh
[1]: https://github.com/mono/mono/pull/12434