[Xamarin.Android.Build.Tests] Don't auto clean up tests. (#2757)
We have a problem in our tests when the build passes, but the unit
test fails.  In this case the `Xamarin.ProjectTools.ProjectBuilder`
type auto cleans up the failed test build directory along with ALL
the logs, which makes it difficult to ascertain why the test failed.

So Lets NOT clean up by default.  This does not mean that we will
leave files on disk for passed tests since the
[`BaseTest.CleanupTest()` method][1] method is an NUnit `[TearDown]`
method.  It runs at the end of the test and will clean up the
directory if the test passed.

[1]: https://github.com/xamarin/xamarin-android/blob/e557aff10ee336b8728ee149d5f2110156a12471/src/Xamarin.Android.Build.Tasks/Tests/Xamarin.Android.Build.Tests/Utilities/BaseTest.cs#L376