[jenkins] fix pipeline publish location (#2766)
Context: https://github.com/xamarin/xamarin-android/pull/2743

Prior to the Xamarin.Android (release) build moving from a freestyle
to a pipeline build, the publish location included both the job name
(xamarin-android) *and* the checkout directory name (xamarin-android)
in the publish url path such as the following - note the redundant
xamarin-android names (`JOB_NAME`/`DIR_NAME`) at the end:

http://xamjenkinsartifact.blob.core.windows.net/mono-jenkins-eval/xamarin-android/xamarin-android

After moving to the pipeline build based on the `build.groovy`
`jenkinsfile`, the publish location changed to the following:

http://xamjenkinsartifact.blob.core.windows.net/mono-jenkins-eval/xamarin-android

This breaks anyone or any system that is looking for published
packages in the original location.

The problem was introduced when the publish operation was
inadvertently changed to execute from within the xamarin-android
checkout directory.  Consequently, the directory itself was not
included in the publish path.

The fix is to blank out the directory specifier and update the list of
relative publish paths such that the publish command runs from the
Jenkins workspace root directory, which is the parent of the
xamarin-android checkout directory.  Consequently, the xamarin-android
checkout directory now gets included in the publish link url.