[tests] Only run AOT tests if the cross compilers exist (#2768)
Context: https://jenkins.mono-project.com/view/Xamarin.Android/job/xamarin-android-pr-builder-release/928/

The AOT-related unit tests require that the corresponding mono
cross-compiler exist.  For example, in order to build
`src/Mono.Android/Test/Mono.Android-Tests.csproj` with
`$(AotAssemblies)`=True -- as is done by the `RunApkTests` target
within `build-tools/scripts/RunTests.targets`, then
`$prefix/lib/xamarin.android/xbuild/Xamarin/Android/cross-arm` must
exist (on macOS) in order for it to build.  If `cross-arm` doesn't
exist, then an XA3001 error will be generated:

	bin/Release/bin/xabuild src/Mono.Android/Test/Mono.Android-Tests.csproj /p:AotAssemblies=True
	...
	error XA3001: System.AggregateException: One or more errors occurred. (ApplicationName='…/cross-arm',
	  CommandLine='--aot=outfile=…,
	  Native error= Cannot find the specified file)

(Aside: the resulting XA3001 error is entirely inscrutable, though
this may not matter as this is a scenario which shouldn't happen.)

What was happening is the following:

  * Commit 241d9416 -- along with external Jenkins changes --
    migrated the xamarin-android Jenkins build to use
    `build-tools/automation/build.groovy`, which had an accidental
    issue in which it would upload the resulting "mono bundle" to the
    wrong directory.

    This in and of itself wasn't problematic, as the existing bundle
    was still valid.

  * The existing mono bundle was invalidated in commit 81f76f83, as
    `src/mono-runtimes/mono-runtimes.projitems` was modified (among
    others), invalidating the bundle hash; see also
    `build-tools/create-bundle/bundle-path.targets`.

After commit 81f76f83 -- and before commit 558901f2, which corrected
the directory that the bundle was uploaded to -- all PR builds would
attempt to download the mono bundle, and that download would fail, as
it didn't exist:

	warning : Unable to download URL `https://…/Release/bundle-v21-h1b15cd2e-Release-Darwin-libzip=b95cf3f,llvm=349752c,mono=b041669.zip` …
	404 (Not Found)

This shouldn't be a problem; it "just" means that the PR builder will
need to rebuild mono.  This will increase PR build times, but
shouldn't otherwise break things...

...except when it does, which is the case here: *By default*,
`make all` only builds a *subset* of mono: JIT runtimes for the host
OS (macOS for the **macOS Release Build** PR builder), armeabi-v7a,
and arm64-v8a.  No AOT compilers are built by `make all`, nor is LLVM
built, nor the Windows JIT, nor...

Which is Fine™, until it isn't: when we execute unit tests which
require that the cross compilers exist.

(Aside: this means that if the PR had the
**full-mono-integration-build** label, the AOT tests would *succeed*,
because the **full-mono-integration-build** label builds *everything*,
not just a faster-to-build subset.)

We *don't* want to make default PR builds build all of mono; that
will increase PR build times when it isn't necessarily warranted.

Instead, make AOT test execution conditional on the AOT
cross-compilers existing.  If the AOT cross-compilers don't exist --
specifically, `cross-arm` (this isn't a thorough check) -- then don't
build the AOT `.apk` tests, which will avoid the XA3001 build errors.

Additionally, remove the MSBuild Condition on
`$(Configuration)`=Release in order to build the AOT tests.  This was
originally done in an effort to reduce PR build times, when the `Debug`
configuration was the default PR configuration, and the "master" job
would build *both* `Debug` *and* `Release` configurations.  Neither
of those points is true anymore; the default PR job builds the
Release configuration, and the "master" job likewise now only builds
the Release configuration.  There is thus no major reason to disable
the AOT tests in the Debug configuration, and *skipping* these tests
could instead *hide* AOT stability issues in the Debug configuration.