[build] Use Homebrew mingw-w64, not MXE (#2758)
MXE, the [M cross environment][mxe], is used to build Windows binary
code on Linux and macOS so that we don't need a Windows machine to
produce our Windows `.vsix` files.  This has many pros and cons,
including:

  * **Pro**: Building *mono* on Windows has historically been
    difficult.  While there is much better support *now* for building
    mono for Windows, we also need to build mono *targeting Android*,
    built via the Android NDK toolchain, and @jonpryor *suspects*,
    but has not verified, that such a cross-compilation scenario
    would still be a terrible experience on Windows.
    
    (This is why Windows builds require a mono bundle; see 29568117.)

  * **Pro**: Building both the macOS and Windows commercial
    installers on macOS means that, where possible, the *same* files
    can be present in both installers.  File such as BCL assemblies,
    `Mono.Android.dll`, MSBuild targets, etc.  Separate macOS &
    Windows build machines could introduce the possibility for
    "drift" and accidental variations between the systems.

  * **Con**: MXE *itself* is a giant pain.  See also: b28856f4, in
    which we see that MXE requires Xcode 8.3 to build; the current
    Xcode version is 10.1, and 8.3 is nearly 2 years old.

      * Xcode 8.3 means we can't use the latest-and-greatest Mac mini
        hardware for builds, as it won't run Xcode 8.3.

      * It can take *tens* of minutes to build MXE...

      * ...assuming it builds AT ALL; it frequently fails to build.
    
Finally, the compilers that MXE provides are *ancient*, providing gcc
5.4.0 (circa 2015), which is 3-4 years old now.  This is crimping
@grendello's style in `src/monodroid`, wherein he wants to use C++17
features when reasonable, but gcc 5.4 doesn't support C++17.  We also
want to use newer compilers on the assumption that they'll generate
more/better warnings around unsafe code practices, improved stack
overflow checking code, and other security-related concerns.

We thus would like to move away from MXE, but to what?

Turns Out™, [Homebrew][hb] now has a [mingw-w64 formula][mingw] to
install the latest MinGW toolchain with GCC 8.2, which allows us to
drop MXE, and we already depend on Homebrew on macOS (82ce2118,
0c073f67f).

The downside of using this pre-packaged MinGW -- a very insignificant
one -- is that it does not come with any extra libraries or  software
prebuilt for Windows, and our build requires `zlib` which was
previously built as part of MXE.  This downside is easily worked
around, however: we simply use our own Homebrew tap which has a
formula that installs the Win32/Win64 `zlib` binary.  This new tap is
available in [xamarin/homebrew-xamarin-android-windeps][windep].

This commit drops MXE dependency in Xamarin.Android *only*; it does
*not* remove this dependency from Mono.  Mono will now provision its
own copy of MXE and use it for builds.  This will change when
[mono/mono#12434][m12434] makes its way into Mono and, subsequently,
to xamarin-android.

All the libraries we build for Windows that require `zlib` will link
a *static* version of the library instead of using the dynamic one.
This is safer than using a `.DLL` because we control what version of
`zlib` is used and we leave no potential for breakage should a 3rd
party replace/remove/etc. `zlib1.dll` from the target system.

[mxe]: http://mxe.cc
[hb]: https://brew.sh
[mingw]: https://formulae.brew.sh/formula/mingw-w64
[m12434]: https://github.com/mono/mono/pull/12434
[windep]: https://github.com/xamarin/homebrew-xamarin-android-windeps