[Xamarin.Android.Build.Tasks] Check for `_._` in the Nuget path.
Context https://devdiv.visualstudio.com/DevDiv/_workitems/edit/679400

When we try to resolve the `runtime` assembly we may well
get a `_._` instead of the filename. This usually means
that the assembly will be provided by the framework. When
this happens we get the following error

	 Error: Exception while loading assemblies: Java.Interop.Tools.Diagnostics.XamarinAndroidException: error XA0009: Error while loading assembly: /Users/tamasszadvari/.nuget/packages/system.runtime.loader/4.3.0/lib/MonoAndroid10/_._

So we should really be ignoring this particular senario.