[build] Use specific ninja binary for mingw cross-builds (#2774)
When Windows cross-build targets are enabled, Xamarin.Android builds
a set of Windows libraries our runtime is dependent upon.  Libraries
are configured with `cmake` and built with `ninja`.  However, the
exact location of the `ninja` binary was not specified when
configuring the libraries for build and, thus, the build may fail if
there's no `ninja` available in the host machine's `$PATH`.

Pass to `cmake` the full path to the Android SDK's `ninja` to make
sure `cmake` finds the binary.