[Xamarin.Android.Build.Tasks] APKs don't need host *.so libraries (#2786)
The `.apk` build process includes a step to put native libraries
(Mono runtime, Xamarin.Android runtime, etc.) into the archive.
This is done by including known `*.so` files from the per-
architecture library directories in the Xamarin.Android source tree
or from a location on the end user's computer.  Unfortunately,
currently this process uses a `lib\*\filename.so` glob pattern to
find the required libraries, and this can include Linux host
libraries with those names, thus producing the following error when
attempting to build any application which needs to embed SQLite,
`libMonoPosixHelper.so` or (in the `mono-2018-10` branch)
`libmono-native.so` shared libraries:

	Adding native library: bin/Debug/lib/xamarin.android/xbuild/Xamarin/Android/lib/arm64-v8a/libmono-profiler-log.so (APK path: lib/arm64-v8a/libmono-profiler-log.so) (TaskId:320)
	  bin/Debug/lib/xamarin.android/xbuild/Xamarin/Android/lib/host-Linux/libmono-native.so : error XA4301: Cannot determine abi of native library
	  bin/Debug/lib/xamarin.android/xbuild/Xamarin/Android/lib/host-Linux/libmono-native.so. [xamarin-forms-samples/FormsGallery/FormsGallery/FormsGallery.Android/FormsGallery.Android.csproj]

Fix the issue by constructing paths to all the supported ABI
locations before attempting to include the shared libraries from
them, thus forgoing the need to use the glob pattern and removing
the problem.