[mingw] Make sure all 32-bit Windows libraries use static libgcc
MinGW builds of shared libraries may (depending on what the source code uses)
link against `libgcc` in a few variations. The default work mode is to use the
shared library but since it's specific to gcc/MinGW, it will not be found on the
target system and the library will fail to load.

There are two solutions for this:

  * Distribute all the referenced DLLs
  * Link them in statically

This commit takes the second approach, implemented in the way that will affect
all the libraries we build for Windows.

Additionally, this commit prepares for the future upgrade to cmake 3.10