[Xamarin.Android.Build.Tasks] Support CustomViews with Aapt2 (#2596)
* [Xamarin.Android.Build.Tasks] Support CustomViews with Aapt2

There is a problem with our `aapt2` support where
custom views do not have their namespaces replaced
with the `{md5}` hash which is generated by
`GenerateJavaStubs`.

This is due to the order in which we need to do certain
operations. For `aapt2` we need to run `aapt2 compile` to
generate the `flata` archives, these are then passed to
`aapt2 link` to generate the `packaged_resources` archive.
The idea was to get the `compile` step done early in
the build so we could reuse the `flata` archives.
We have to do this for generating the `Resources.desinger.cs`.

The problem is `GenerateJavaStubs` does not run until AFTER
`CoreCompile` at this point the `flata` archives are already
in place. As a result they do not include the fixed up
custom views. So to work around this issue we need to run
`aapt2 compile` after `GenerateJavaStubs` for those items
which contain custom views.

* rework a bit

* Add build with no changes check

* Fix test

* Update ConvertCustomView.cs

Clean up `resdir?`, since there is now a null check.