[Xamarin.Android.Build.Tasks] Remove usage of Before/AfterTargets (#2795)
Fixes: https://github.com/xamarin/xamarin-android/issues/1128
Context: https://github.com/jonathanpeppers/MSBuildIncrementalClean

Through research into MSBuild, there are two problem we generally
want to fix:

 1. We have use of `AfterTargets` and `BeforeTargets`, which causes
    some problems around error handling.
 2. We have targets we want to run *before* `IncrementalClean`.
    Our current implementation uses a private MSBuild target to
    achieve this.

~~ AfterTargets & BeforeTargets ~~

The problem with `AfterTargets` and `BeforeTargets` is around error
handling: what should happen when a "previous" target generates an
error?  Should the subsequent targets execute or not?

When `AfterTargets` and `BeforeTargets` are used, errors from
"previous targets" are *ignored*.  For example, via `BeforeTargets`
and `DependsOnTargets`, the `_LinkAssemblies` target executes *after*
the `Compile` target and before the `IncrementalClean` target.

What happens if the `Compile` target fails?

	MainActivity.cs(9,7): error CS0246: The type or namespace name 'ClassLibrary1' could not be found (are you missing a using directive or an assembly reference?)

Doesn't matter what the reason is; it's user code, it should be
expected to fail occasionally.  This should be fine.

What *isn't* fine is that after the `Compile` target files, the
`_LinkAssemblies` target fails!

	C:\Program Files (x86)\Microsoft Visual Studio\2019\Enterprise\MSBuild\Xamarin\Android\Xamarin.Android.Common.targets(2160,5): error MSB4018: The "LinkAssemblies" task failed unexpectedly.
	System.IO.FileNotFoundException: Could not load assembly 'App1, Version=0.0.0.0, Culture=neutral, PublicKeyToken='. Perhaps it doesn't exist in the Mono for Android profile?
	    File name: 'App1.dll'
	    at Java.Interop.Tools.Cecil.DirectoryAssemblyResolver.Resolve(AssemblyNameReference reference, ReaderParameters parameters)
	    at Java.Interop.Tools.Cecil.DirectoryAssemblyResolver.GetAssembly(String fileName)
	    at Xamarin.Android.Tasks.LinkAssemblies.Execute(DirectoryAssemblyResolver res)
	    at Xamarin.Android.Tasks.LinkAssemblies.Execute()
	    at Microsoft.Build.BackEnd.TaskExecutionHost.Microsoft.Build.BackEnd.ITaskExecutionHost.Execute()
	    at Microsoft.Build.BackEnd.TaskBuilder.<ExecuteInstantiatedTask>d__26.MoveNext() [C:\Users\joaqu\source\repos\ClassLibrary21\App1\App1.csproj]

The `<LinkAssemblies/>` task fails because one of the input
assemblies, `App1.dll`, doesn't exist.  It doesn't exist *because*
the `Compile` target failed.

The `<LinkAssemblies/>` task *shouldn't even execute* in this
situation; it cannot possibly succeed if the `Compile` target fails.
Worse, the `<LinkAssemblies/>` errors are displayed in the IDE!

The only way to prevent the `<LinkAssemblies/>` task from executing
if the `Compile` target fails is to *avoid* use of `AfterTargets` and
`BeforeTargets`.  Instead, properties such as `$(CoreBuildDependsOn)`
need to be overridden, as in *that* context MSBuild will stop
executing targets as soon as a target fails.

Going through our targets here, there was a decent bit of rework
needed. I also updated our MSBuild "best practices" docs.


~~ IncrementalClean ~~

`IncrementalClean` is in a similar boat: we need the
`_AddFilesToFileWrites` target to execute *before* `IncrementalClean`
executes so that `@(FileWrites)` is correct, and because
`_AddFilesToFileWrites` used `BeforeTargets`, it couldn't participate
"normally" in target error behavior, i.e. it would *always* execute.

We thus don't want to use `BeforeTargets`.

However, there is no `$(IncrementalCleanDependsOn)` property to
override, which raises an interesting conundrum: how do we get
targets to execute *before* `IncrementalClean` *without* using
`BeforeTargets`?

The answer is by using the [`$([MSBuild]::Unescape())`][msbu]
property function (thanks to @rainersigwald with this!):

> `string Unescape(string escaped)`  
> Unescape the string according to MSBuild escaping rules.

[msbu]: https://docs.microsoft.com/en-us/visualstudio/msbuild/property-functions?view=vs-2017#msbuild-property-functions

	<PropertyGroup>
	  <!--Add to this property as needed here-->
	  <_BeforeIncrementalClean>
	    _AddFilesToFileWrites;
	  </_BeforeIncrementalClean>
	  <CoreBuildDependsOn>
	    $([MSBuild]::Unescape($(CoreBuildDependsOn.Replace('IncrementalClean;', '$(_BeforeIncrementalClean);IncrementalClean;'))))
	  </CoreBuildDependsOn>
	</PropertyGroup>

A "normal" `CoreBuildDependsOn.Replace(...)` invocation would leave
`;` "escaped", meaning nothing would work as desired.  Using the
`Unescape()` property function ensures that the `;` *isn't* escape,
which in turn allows the resulting `$(CoreBuildDependsOn)` to have
the intended semantics: a `;`-separated list of targets.

This voodoo allows us to specify targets that need to run before the
`IncrementalClean` target, without using `BeforeTargets`.

This in turn allows removing the `BeforeTargets` use within
`_CompileDex` and `_PrepareAssemblies`.


~~ _SetupApplicationJavaClass ~~

The `_SetupApplicationJavaClass` target had:

	<Target Name="_SetupApplicationJavaClass"
	    AfterTargets="_ResolveMonoAndroidSdks"
	    DependsOnTargets="$(_BeforeSetupApplicationJavaClass)">

I moved the contents of this target to `_ResolveMonoAndroidSdks`.

Usage of `$(_BeforeSetupApplicationJavaClass)` will need to change to
use `$(_ResolveMonoAndroidSdksDependsOn)` instead.


~~ _ResolveMonoAndroidFramework ~~

The `_ResolveMonoAndroidFramework` target was empty, and I could not
find anything using it.

	<Target Name="_ResolveMonoAndroidFramework" DependsOnTargets="GetReferenceAssemblyPaths" />

I removed it, and used `GetReferenceAssemblyPaths` where applicable.


~~ _SetLatestTargetFrameworkVersion ~~

The `_CreateAapt2VersionCache` target was using
`AfterTargets="_SetLatestTargetFrameworkVersion"`, so this target
needed to be reworked.

I renamed `_SetLatestTargetFrameworkVersion` to `_ResolveSdks`.

I created a new, empty `_SetLatestTargetFrameworkVersion` target that
depends on:

  * `_ResolveSdks`, then
  * `_CreateAapt2VersionCache`


~~ Changes needed in monodroid ~~

  * Usage of `$(_BeforeSetupApplicationJavaClass)` will need to use
    `$(_ResolveMonoAndroidSdksDependsOn)`
  * Need to do a general audit for `BeforeTargets` & `AfterTargets`