Revert "[BootstrapTasks] libZipSharp & <Unzip…/> (#2750)" (#2796)
This reverts commit e84e502e32b7a88f3cfd46a0b3bd6bcbb5f8fba4.

While LibZipSharp gives us better performance, we have a chicken-and-
egg problem: in order to work libZipSharp needs the native `libzip`
library, but `libZipSharp` has to be built by `xamarin-android`
before we can use it.  This creates a problem when bootstrapping
Xamarin.Android either on a completely clean machine (i.e. without
`$HOME/android-toolchain`) or when updating one of the Android
SDK/NDK components required for build.  In such instances
`make prepare` will fail because `libzip` needs to be built using
`cmake` which isn't yet installed and so the preparation step fails.

Revert commit e84e502e in order to get the old behavior back until we
can fix the original problem fixed by e84e502e in a way that doesn't
create the above problems.