[build] Retry some steps for better network resiliency (#2797)
Context: https://jenkins.mono-project.com/job/xamarin-android-pr-pipeline-release/30

With the xamarin-android build being applied to PR builds, it is more
critical that the build be more resilient to connectivity glitches.
There is one such case where the build failed during the checkout
stage due to known network issues.  There is a good chance that the
checkout stage would have succeeded if given another chance:

	 > git submodule update --init --recursive external/llvm
	...
	ERROR : checkout: Unexpected error: java.io.IOException: Could not perform submodule update. Marking build as ERROR.

Add retry capability to the build.groovy pipeline script, which
includes a wait between retries.  The wait increases exponentially on
each retry.  Retries have been enabled for stages that depend on
connectivity for checkout and publishing of results.  These stages
are good candidates for retry as they will most likely fail due to
connectivity issues as opposed to an issue with building the code.

Additionally, bumped up the timeouts for the publishing stages as
they involve publishing of build logs that in some situations can
turn out to be quite large, e.g. a recent mono/2018-10 bump generates
a ~1.5GB log file.

Finally, incorporates a fix for a rogue use of a `stageStatus`
variable.  Updated all stages to align on using `commandStatus` as
the variable name.