Xamarin.Android build preparation utility
Testing:

 * Default mode: run `make prepare`
 * Verbose debug: run `make V=1 prepare`
 * CI mode: run `make PREPARE_CI=1 prepare`
 * Help: run `make prepare-help`
 * Logs: `bin/BuildDebug`

Supported operating systems:

 * macOS
 * Linux

Windows support is planned, present but not fully implemented or tested.

This commit implements a `make prepare` replacement that does not rely on shell
scripts or MSBuild. The idea is that a standalone C# program can perform all the
steps in a more streamlined, more clear way than the current solution. One of
the main goals is to make the process more approachable by external
contributors, but also by the core developers who may want to change some
aspects of XA build preparation (e.g. Android platforms, Mono version etc) but
are not familiar with the build system. Towards that goal, the codebase is
designed so that the minimum amount of searching around it is necessary to
figure out where to make the desired change. Main location serving this purpose
is the `build-tools/xaprepare/xaprepare/ConfigAndData` directory. It should
be the *only* location where one needs to look in order to change all and any
aspects of XA preparation.

[TBC]